third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/stylesheet-source-url-comment.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<style>
body {
    color: green;
}

/*# sourceURL=inlineStyleSheet.css */
</style>
<script>
function addInlineStyleSheet()
{
    var styleElement = document.createElement("style");
    styleElement.textContent = "body { color: black; }\n/*# sourceURL=css/addedInlineStylesheet.css */";
    document.head.appendChild(styleElement);
}

function addInlineStyleSheetNonRelative()
{
    var styleElement = document.createElement("style");
    styleElement.textContent = "body { color: red; }\n/*# sourceURL=css/nonRelativeInlineStylesheet.css */";
    document.head.appendChild(styleElement);
}

function addInlineStyleSheetMultiple()
{
    var styleElement = document.createElement("style");
    styleElement.textContent = "\n/*# sourceURL=1.css */\nbody { color: red; }\n/*# sourceURL=2.css*/\n/*# sourceURL=css/addedInlineStylesheetMultiple.css */";
    document.head.appendChild(styleElement);
}

function addInlineStyleSheetDeprecated()
{
    var styleElement = document.createElement("style");
    styleElement.textContent = "body { color: black; }\n/*@ sourceURL=css/addedInlineStylesheetDeprecated.css */";
    document.head.appendChild(styleElement);
}

function test() {
  function forEachHeaderMatchingURL(url, handler) {
    var headers = TestRunner.cssModel.styleSheetHeaders();
    for (var i = 0; i < headers.length; ++i) {
      if (headers[i].sourceURL.indexOf(url) !== -1)
        handler(headers[i]);
    }
  }

  function checkHeaderHasSourceURL(header) {
    TestRunner.assertTrue(header.hasSourceURL);
  }

  TestRunner.runTestSuite([
    function testSourceURLCommentInInlineScript(next) {
      SourcesTestRunner.showScriptSource('stylesheet-source-url-comment.html', didShowSource);

      function didShowSource(sourceFrame) {
        function checkHeaderDoesNotHaveSourceURL(header) {
          TestRunner.assertTrue(!header.hasSourceURL, 'hasSourceURL flag is set for inline stylesheet');
        }

        forEachHeaderMatchingURL('source-url-comment.html', checkHeaderDoesNotHaveSourceURL);
        next();
      }
    },

    function testSourceURLComment(next) {
      SourcesTestRunner.showScriptSource('css/addedInlineStylesheet.css', didShowSource);
      TestRunner.evaluateInPage('setTimeout(addInlineStyleSheet, 0)');

      function didShowSource(sourceFrame) {
        TestRunner.addResult(sourceFrame.textEditor.text());
        forEachHeaderMatchingURL('addedInlineStylesheet', checkHeaderHasSourceURL);
        next();
      }
    },

    function testDeprecatedSourceURLComment(next) {
      SourcesTestRunner.showScriptSource('css/addedInlineStylesheetDeprecated.css', didShowSource);
      TestRunner.evaluateInPage('setTimeout(addInlineStyleSheetDeprecated, 0)');

      function didShowSource(sourceFrame) {
        TestRunner.addResult(sourceFrame.textEditor.text());
        forEachHeaderMatchingURL('addedInlineStylesheetDeprecated', checkHeaderHasSourceURL);
        next();
      }
    },

    function testNonRelativeURL(next) {
      SourcesTestRunner.showScriptSource('css/nonRelativeInlineStylesheet.css', didShowSource);
      TestRunner.evaluateInPage('setTimeout(addInlineStyleSheetNonRelative, 0)');

      function didShowSource(sourceFrame) {
        TestRunner.addResult(sourceFrame.textEditor.text());
        forEachHeaderMatchingURL('nonRelativeInlineStyleSheet.css', checkHeaderHasSourceURL);
        next();
      }
    },

    function testMultiple(next) {
      SourcesTestRunner.showScriptSource('css/addedInlineStylesheetMultiple.css', didShowSource);
      TestRunner.evaluateInPage('setTimeout(addInlineStyleSheetMultiple, 0)');

      function didShowSource(sourceFrame) {
        TestRunner.addResult(sourceFrame.textEditor.text());
        forEachHeaderMatchingURL('addInlineStyleSheetMultiple.css', checkHeaderHasSourceURL);
        next();
      }
    }
  ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests that stylesheets with sourceURL comment are shown in the Sources panel.</p>
</body>
</html>
