third_party/blink/web_tests/http/tests/devtools/elements/styles-4/stylesheet-source-url-comment-expected.txt

Tests that stylesheets with sourceURL comment are shown in the Sources panel.


Running: testSourceURLCommentInInlineScript

Running: testSourceURLComment
body { color: black; }

Running: testDeprecatedSourceURLComment
body { color: black; }

Running: testNonRelativeURL
body { color: red; }

Running: testMultiple

/*# sourceURL=1.css */
body { color: red; }
/*# sourceURL=2.css*/
